+++ backend/app/services/import_execution_summary.py
@@ -1,0 +1,56 @@
+"""Job summary and audit helpers for Workbench import execution."""
+
+from __future__ import annotations
+
+import uuid
+from typing import Any, Literal
+
+from sqlmodel import Session
+
+from app.models import User
+from app.models.base import get_datetime_utc
+from app.services.audit import record_audit_event
+
+
+def _record_import_audit(
+    session: Session,
+    *,
+    current_user: User,
+    project_id: uuid.UUID,
+    run_id: uuid.UUID,
+    status: Literal["success", "failure"],
+    stage: str,
+    input_type: str,
+) -> None:
+    record_audit_event(
+        session,
+        action="import.run",
+        resource_type="analysis_run",
+        resource_id=run_id,
+        status=status,
+        actor=current_user,
+        project_id=project_id,
+        detail={"stage": stage, "input_type": input_type},
+    )
+
+
+def _job_payload(
+    *,
+    job_id: str,
+    status: str,
+    status_history: list[dict[str, str]],
+) -> dict[str, Any]:
+    timestamp = get_datetime_utc().isoformat()
+    return {
+        "id": job_id,
+        "status": status,
+        "updated_at": timestamp,
+        "status_history": status_history,
+    }
+
+
+def _job_status_entry(status: str) -> dict[str, str]:
+    return {
+        "status": status,
+        "created_at": get_datetime_utc().isoformat(),
+    }
