+++ backend/tests/test_refactor_hygiene.py
@@ -501,13 +501,32 @@ def test_backend_package_boundary_intentionally_ships_workbench_app() -> None:
     assert "package-check: package-contents-check" in makefile
 
 
-def test_import_execution_has_characterization_guardrail_before_deeper_split() -> None:
+def test_import_execution_is_split_into_stage_services_with_guardrails() -> None:
     source = (ROOT / "app/services/import_execution.py").read_text(encoding="utf-8")
+    context_source = (ROOT / "app/services/import_execution_context.py").read_text(encoding="utf-8")
+    failure_source = (ROOT / "app/services/import_execution_failures.py").read_text(
+        encoding="utf-8"
+    )
+    persistence_source = (ROOT / "app/services/import_execution_persistence.py").read_text(
+        encoding="utf-8"
+    )
+    summary_source = (ROOT / "app/services/import_execution_summary.py").read_text(encoding="utf-8")
 
     assert "def execute_project_import_upload" in source
-    assert "def _apply_template_asset_context" in source
-    assert "def _apply_template_vex" in source
-    assert "def _persist_template_occurrences" in source
-    assert "def _job_payload" in source
-    assert "def _parse_error_payload" in source
-    assert len(source.splitlines()) <= 1800
+    assert "app.services.import_execution_context" in source
+    assert "app.services.import_execution_failures" in source
+    assert "app.services.import_execution_persistence" in source
+    assert "app.services.import_execution_summary" in source
+    assert "def _apply_template_asset_context" in context_source
+    assert "def _apply_template_vex" in context_source
+    assert "def _parse_error_payload" in context_source
+    assert "def raise_analysis_failure" in failure_source
+    assert "def _persist_template_occurrences" in persistence_source
+    assert "def _persist_template_occurrences_bulk_insert" in persistence_source
+    assert "def _job_payload" in summary_source
+    assert "def _record_import_audit" in summary_source
+    assert len(source.splitlines()) <= 700
+    assert len(context_source.splitlines()) <= 220
+    assert len(failure_source.splitlines()) <= 140
+    assert len(persistence_source.splitlines()) <= 1000
+    assert len(summary_source.splitlines()) <= 100
